fix: validando campos de edição
