+++ src/pages/assistentes/assistentes.tsx
@@ -1,9 +1,8 @@
-import React, { useState } from "react";
-import Sidebar from "../../shared/components/Sidebar/sidebar";
+import { useState } from "react";
 import Navbarlog from "../../shared/components/NavbarLogada/navbarLogada";
-import DataTable from "../../shared/components/TablePagination/tablePagination";
 import PrimaryButton from "../../shared/components/PrimaryButton/PrimaryButton";
-import * as EmailValidator from 'email-validator';
+import Sidebar from "../../shared/components/Sidebar/sidebar";
+import DataTable from "../../shared/components/TablePagination/tablePagination";
 
 import {
   Box,
@@ -19,11 +18,11 @@ import {
   TextField,
   Typography,
 } from "@mui/material";
-import { useQuery } from "react-query";
-import { FormProvider, useForm } from "react-hook-form";
 import { GridActionsCellItem, GridRowId } from "@mui/x-data-grid";
-import { BsFillTrashFill } from "react-icons/bs";
+import { FormProvider, useForm } from "react-hook-form";
 import { AiFillEdit, AiFillEye, AiFillEyeInvisible } from "react-icons/ai";
+import { BsFillTrashFill } from "react-icons/bs";
+import { useQuery } from "react-query";
 import { toast } from "react-toastify";
 import {
   cadastrarAssistente,
@@ -31,12 +30,17 @@ import {
   excluirAssistente,
   listarAssistentes,
 } from "../../services/assistentes";
-import { AssistentesCadastrarDTO } from "./dtos/AssistentesCadastrar.dto";
-import { AssistentesListarDTO } from "./dtos/AssistentesListar.dto";
 import { queryClient } from "../../services/queryClient";
 import ValueMask from "../../shared/components/Masks/ValueMask";
-import validarCPF from "../../shared/functions/validarCPF";
-import removeSpecialCharacters from "../../shared/functions/removeSpecialCharacters";
+import removeSpecialCharacters from "../../shared/validations/removeSpecialCharacters";
+import { validateCPF } from "../../shared/validations/validarCPF";
+import { validateAge, validateDate } from "../../shared/validations/validarDataNascimento";
+import { validateEmail } from "../../shared/validations/validarEmail";
+import { validateLogin } from "../../shared/validations/validarLogin";
+import { validateNome } from "../../shared/validations/validarNome";
+import { validateSenha } from "../../shared/validations/validarSenha";
+import { AssistentesCadastrarDTO } from "./dtos/AssistentesCadastrar.dto";
+import { AssistentesListarDTO } from "./dtos/AssistentesListar.dto";
 
 import {
   getContainerStyles,
@@ -83,93 +87,6 @@ export function Assistentes() {
     formState: { errors },
   } = methods;
 
-  const validateCPF = (cpf: string): boolean => {
-
-    const cpfEhValido = validarCPF(cpf);
-    if (!cpfEhValido) {
-      toast.error("O CPF informado é inválido.");
-      return false;
-    }
-    return true;
-  };
-
-  const validateEmail = (email: string): boolean => {
-    const emailValido = EmailValidator.validate(email);
-    if (!emailValido) {
-      toast.error("O e-mail informado é inválido.");
-      return false;
-    }
-    return true;
-  };
-
-  const validateDate = (date: string): boolean => {
-    const dateRegex = /^(\d{2})\/(\d{2})\/(\d{4})$/;
-    if (!dateRegex.test(date)) {
-      toast.error("Formato de data inválido. Use o formato dd/mm/aaaa.");
-      return false;
-    }
-    const matchResult = dateRegex.exec(date);
-
-    if (!matchResult) {
-      toast.error("Data de nascimento inválida.");
-      return false;
-    }
-    const [, dia, mes, ano] = matchResult;
-    const dataNascimento = new Date(Number(ano), Number(mes) - 1, Number(dia));
-    if (
-      dataNascimento.getFullYear() !== Number(ano) ||
-      dataNascimento.getMonth() !== Number(mes) - 1 ||
-      dataNascimento.getDate() !== Number(dia)
-    ) {
-      toast.error("Data de nascimento inválida.");
-      return false;
-    }
-    return true;
-  };
-
-  const validateNome = (nome: string): boolean => {
-    if (nome.length > 70) {
-      toast.error("Nome inválido.");
-      return false;
-    }
-    return true;
-  };
-
-  const validateLogin = (login: string): boolean => {
-    if (login.length < 8) {
-      toast.error("Login muito pequeno.");
-      return false;
-    }
-    return true;
-  };
-
-  const validateSenha = (senha: string): boolean => {
-    if (senha.length < 8) {
-      toast.error("Senha muito pequena.");
-      return false;
-    }
-    return true;
-  };
-
-  const validateAge = (dataNascimento: Date): boolean => {
-    const hoje = new Date();
-    let idade = hoje.getFullYear() - dataNascimento.getFullYear();
-    const mesAtual = hoje.getMonth();
-    const mesNascimento = dataNascimento.getMonth();
-    if (
-      mesAtual < mesNascimento ||
-      (mesAtual === mesNascimento && hoje.getDate() < dataNascimento.getDate())
-    ) {
-      idade--;
-    }
-    if (idade < 18) {
-      toast.error("É necessário ter mais de 18 anos para este cadastro.");
-      return false;
-    }
-    return true;
-  };
-
-
   const registerAssistentes = async (data: any) => {
     const assistente = {
       nome: data.nome,
@@ -183,6 +100,8 @@ export function Assistentes() {
       observacao: data.observacao,
       administrador: true,
     } as AssistentesCadastrarDTO;
+
+    // VALIDAÇÕES
 
     const cpfValido = validateCPF(assistente.cpf);
     if (!cpfValido) {
@@ -231,7 +150,7 @@ export function Assistentes() {
     if (!senhaValida) {
       return;
     }
-
+    
     assistente.cpf = removeSpecialCharacters(assistente.cpf);
     assistente.telefone = removeSpecialCharacters(assistente.telefone);
     assistente.dNascimento = transformDate(assistente.dNascimento);
@@ -283,10 +202,10 @@ export function Assistentes() {
   });
 
   const deleteAssistentes = async () => {
-    const selectedAssistente = dataTable.find((item) => (item as any).id === id); // Encontra o objeto da aluna com base no ID selecionado
+    const selectedAssistente = dataTable.find((item) => (item as any).id === id); // Encontra o objeto da assistente com base no ID selecionado
     if (selectedAssistente) {
-      const login = (selectedAssistente as any).login; // Obtém o login da aluna
-      const response = await excluirAssistente(login); // Passa o login para a função apagaAluna
+      const login = (selectedAssistente as any).login; // Obtém o login da assistente
+      const response = await excluirAssistente(login); // Passa o login para a função apagaassistente
 
       if (response.status === 204) {
         toast.success("Assistente excluída com sucesso!");
@@ -314,15 +233,56 @@ export function Assistentes() {
     setValue("data_nascimentoEdit", assistente.dNascimento);
     setValue("telefoneEdit", assistente.telefone);
     setValue("emailEdit", assistente.email);
-    setValue("observacaoEdit",assistente.login);
+    setValue("observacaoEdit",assistente.observacao);
     setOpenEdit(true);
   };
 
   const editAssistentes = async (data: any) => {
-    data.data_nascimentoEdit = transformDate(data.data_nascimentoEdit);
+    // VALIDAÇÕES
+
+    const cpfValido = validateCPF(data.cpfEdit);
+    if (!cpfValido) {
+      return;
+    }
+
+    const emailValido = validateEmail(data.emailEdit);
+    if (!emailValido) {
+      return;
+    }
+
+    const dateValido = validateDate(data.data_nascimentoEdit);
+    if (!dateValido) {
+      return;
+    }
+
+    const dateRegex = /^(\d{2})\/(\d{2})\/(\d{4})$/;
+
+    const matchResult = dateRegex.exec(data.data_nascimentoEdit);
+
+    if (!matchResult) {
+      toast.error("Data de nascimento inválida.");
+      return;
+    }
+    const [, dia, mes, ano] = matchResult;
+
+    const dataNascimento = new Date(Number(ano), Number(mes) - 1, Number(dia));
+
+    const ageValida = validateAge(dataNascimento);
+    if (!ageValida) {
+      return;
+    }
+
+    const nomeValido = validateNome(data.nomeEdit);
+    if (!nomeValido) {
+      return;
+    }
+
+
     data.cpfEdit = removeSpecialCharacters(data.cpfEdit);
     data.telefoneEdit = removeSpecialCharacters(data.telefoneEdit);
     data.cepEdit = removeSpecialCharacters(data.cepEdit);
+    data.data_nascimentoEdit = transformDate(data.data_nascimentoEdit);
+
 
     const assistenteEditada = {
       nome: data.nomeEdit,
