+++ src/shared/validations/removeSpecialCharacters.tsx
